[3.4] REFACTOR -> class name change
